모듈화 - saveProfile, processCourseData 모듈 단위로 분리
