+++ src/utils/index.js
@@ -1,0 +1,4 @@
+import saveProfile from './saveProfile'
+import processCourseData from './processCourseData'
+
+export { saveProfile, processCourseData }
